Rearranging types
