+++ react/Gallery.tsx
@@ -1,7 +1,27 @@
 import React from 'react'
 import GalleryLayout from './GalleryLayout'
+import { GalleryLayoutProps, Slots } from './GalleryLayout'
 //@ts-ignore
 import GalleryLegacy from './GalleryLegacy'
+
+interface GalleryLegacyProps {
+  lazyItemsRemaining: number
+  maxItemsPerRow: number
+  minItemWidth: number
+  mobileLayoutMode: MobileLayoutMode
+  products: Product[]
+  showingFacets: boolean
+  summary: unknown
+}
+
+type MobileLayoutMode = 'normal' | 'small' | 'inline'
+
+/*
+ * This type receives Slots directly, instead of using the 'slots' prop to do it.
+ * Is the equivalent of adding '[key: string]: ComponentType' at the end of GalleryLayoutProps
+ * and removing the 'slots' prop.
+ */
+type GalleryLayoutPropsWithSlots = Omit<GalleryLayoutProps, 'slots'> & Slots
 
 const Gallery: React.FC<
   GalleryLegacyProps | GalleryLayoutPropsWithSlots
@@ -29,6 +49,59 @@ const Gallery: React.FC<
   }
 
   return <GalleryLegacy {...(props as GalleryLegacyProps)} />
+}
+
+export interface Product {
+  /** Product's id. */
+  productId: string
+  /** Product's name. */
+  productName: string
+  /** Product's description. */
+  description: string
+  /** Product's categories. */
+  categories: unknown[]
+  /** Product's link. */
+  link?: string
+  /** Product's link text. */
+  linkText: string
+  /** Product's brand. */
+  brand?: string
+  /** Product's SKU items. */
+  items: ProductItem[]
+}
+
+interface ProductItemReference {
+  Value: string
+}
+
+interface ProductItemImage {
+  /** Images's imageUrl. */
+  imageUrl: string
+  /** Images's imageTag. */
+  imageTag: string
+}
+
+interface ProductItemSeller {
+  /** Sellers's commertialOffer. */
+  commertialOffer: {
+    /** CommertialOffer's price. */
+    Price: number
+    /** CommertialOffer's list price. */
+    ListPrice: number
+  }
+}
+
+interface ProductItem {
+  /** SKU's id. */
+  itemId: string
+  /** SKU's name. */
+  name: string
+  /** SKU's referenceId. */
+  referenceId?: ProductItemReference[]
+  /** SKU's images. */
+  images: ProductItemImage[]
+  /** SKU's sellers. */
+  sellers: ProductItemSeller[]
 }
 
 export default Gallery
